fix minor issues

+++ src/app/modules/Comment/comment.service.ts
@@ -1,24 +1,43 @@
+import ApiError from '../../../errors/ApiError';
 import prisma from '../../../shared/prisma';
+import httpStatus from 'http-status'
 
-export const handlePostComment = async (providerId: string, receiverId: string, text: string, replyTo?:string ) => {
+export const handlePostComment = async (providerId: string,  text: string,photoId?: string, replyTo?:string ) => {
+    
 
     if(replyTo){
+        const parentComment = await prisma.comment.findUnique({where:{id:replyTo}})
+        if(!parentComment){
+            throw new ApiError(httpStatus.NOT_FOUND, "comment not found to reply")
+        }
         const comment = await prisma.comment.create({
-            data: { providerId, receiverId, text,parentId:replyTo}
+            data: { providerId,  text,parentId:replyTo},
+            include:{provider:{select:{avatar:true, fullName:true, firstName:true, lastName:true}}}
         });
         return comment
     }
+    if(!photoId){
+        throw new ApiError(httpStatus.BAD_REQUEST, "photo id is required")
+    }
+     const photo = await prisma.userPhoto.findUnique({where:{id:photoId}})
+
+    if(!photo){
+        throw new ApiError(httpStatus.NOT_FOUND, "photo not found")
+    }
     const comment = await prisma.comment.create({
-        data: { providerId, receiverId, text }
+        data: { providerId, photoId, text },
+        include:{provider:{select:{avatar:true, fullName:true, firstName:true, lastName:true}}}
     });
 
     return comment;
 };
 
-export const handleGetUserComments = async (userId: string) => {
+export const handleGetUserComments = async (photoId: string) => {
     const comments = await prisma.comment.findMany({
-        where: { receiverId: userId },
-        include: { provider: true ,CommentReplies:true}
+        where: { photoId },
+        include: { provider: {select:{avatar:true, fullName:true, firstName:true, lastName:true}} ,commentReplies:
+    {include:{commentReplies:{include:{provider:{select:{avatar:true, firstName:true,lastName:true, fullName:true}}}}, provider:{select:{avatar:true, fullName:true, firstName:true, lastName:true}}},}},
+        orderBy:{createdAt:"desc"}
     });
     
 
@@ -26,13 +45,17 @@ export const handleGetUserComments = async (userId: string) => {
 };
 
 
-export const getAll = async (receiverId:string)=>{
-    const comments = await prisma.comment.findMany({where:{receiverId}})
+export const getAll = async (photoId:string)=>{
+    const comments = await prisma.comment.findMany({where:{photoId}})
 
     return comments
 }
 
-export const handlDeleteComment = async (commentId:string)=>{
+export const handlDeleteComment = async (userId:string,commentId:string)=>{
+    const comment = await prisma.comment.findUnique({where:{id:commentId}})
+    if(!comment || comment.providerId !== userId){
+        throw new ApiError(httpStatus.BAD_REQUEST, "unable to delete the comment")
+    }
     const deletedComment = await prisma.comment.delete({where:{id:commentId}})
 
     return deletedComment
